:package: relocated custom hooks to hooks subdirectory
